Update patch set 1
Patch Set 1:

No worries Pam, thank you, trying to get to some conclusion, will try to revert to an older image in nexus3 and retry for comparison.   In the integration lab, it is using Alpine images (5 days old) and processes come up OK on the PAP side, but there are some other issues as I can see.

Patch-set: 1