Update patch set 1
Patch Set 1:

Tracking problem under POLICY-1604.  alpine openjdk8 deploys with a smaller set of libraries than previous version.   The one that is causing these problems are libnss3.so in particular.   It seems that the openjdk8 version was released 2 days ago, which explains that we are seeing these problems now.

Patch-set: 1