Update patch set 1
Patch Set 1: Code-Review+2

And in case you are wondering, the 5.0.0 *was* for the El Alto Early Drop - which never happened. But because some projects released 5.0.0 and used that tag on their repos, ONAP overall needs to bump to 5.0.1 for every project to be consistent.

Patch-set: 1
Reviewer: Gerrit User 57 <57@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+2