Override oparent sonar
We don't have any IT tests, so let's skip them. This
may have a trivial bump in speed.

Use the same datafile, and make sure the results are
being appended to support multi-module.

This will result in all the repos removing this
configuration as it only needs to be set here in
policy/parent. Cleans up things a bit.

Issue-ID: POLICY-2321
Change-Id: I999cb3bb8a7a764ffc7459fc9e0b9b8377dbecb3
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>