Update patch set 1
Patch Set 1:

May be, but this resolves one of the issue with :ref: "Installing or Upgrading Policy" as a label used and we have a reference of it on the page. Which it wasn't picking without this.
 
As per  https://sublime-and-sphinx-guide.readthedocs.io/en/latest/references.html
Allow reference sections using its title.

I have tested it, a) it removed the warning. b)link become accessible on the page for "Installing or Upgrading Policy" referencing the section.

Patch-set: 1