Correct documentation for apex installation test
The documentation for the installation test of apex specified that
is changed to use the apexApps.sh script rather than the apexEngine.sh script,
the apexApps.sh script runs apex with the correct userid of "apexuser".

The installation test documentation is corrected to use the -m parameter to point
out the model, otherwise the user has to be in $APEX_HOME for the example to run
the test.

Issue-ID: POLICY-2791
Change-Id: I4d18b5e2c86645183ae65619adf664e7adfefebc
Signed-off-by: liamfallon <liam.fallon@est.tech>