Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> Don't think there were any real changes since we reverted the Jackson - but we can release it if you want.

So I discovered - I originally thought we had pointed at a new version of oparent, but that was actually done in version 3.2.0.  Any point to continue this review?  Or just abandon it for now and wait for more documentation to come in?

Patch-set: 1
Reviewer: Gerrit User 811 <811@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>