Remove css files from apex-pdp s3p results
The Javascript rendering of the results is removed and replaced with PDF
files in the apex_s3p_results.zip file.

The directory structure is changed to remove the reference to the
Frankfurt release.

The link in the rst file is updated to take account of the
changed directory strucuture.

Issue-ID: POLICY-3002
Change-Id: I1bd3ca46107179059de5167923c44f215796fce4
Signed-off-by: liamfallon <liam.fallon@est.tech>