Update patch set 1
Patch Set 1:

It seems that the problems in Michael's lab were resource related.  It run OK with pretty much the default configuration in windriver.  The only change I made was to turn off logging to be on the safe side.   I removed the performance section in case we don't have time to cover it in this release.

Patch-set: 1