Update patch set 1
Patch Set 1:

I have doubts if there is a lot of value including the zip files as the results will vary somewhat from configuration to configuration, and the size of the cloned repos.   We don't include them in some policy projects.   They could be kept in the wiki as an alternative.

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>