Update patch set 3
Patch Set 3: Code-Review+2

OK, I'm not comfortable with ceding so much control to Springboot. I think we should try and pick the individual dependencies we need rather than going "all in" on Springboot for our dependencies. If for example, the LF decide that something that we drag in using Springboot, we would be in bother.

Patch-set: 3
Reviewer: Gerrit User 2787 <2787@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+2