Update parent pom to fix springboot
The spring boot dependency management is added.

The version of Spring in oparent (5.2.10.RELEASE) does not work
with Springboot 2.5.1, it need a newer version of Spring.

This patch (temporarily) overrides the spring dependencies in
oparent until the patch in oparent to do the upgrade is
merged. That patch may take time to get merged because the
chane could impact other projects.

Issue-ID: POLICY-3209
Change-Id: Ie9ab5b001ed9636fa5b4d96f2d54a7e3e800fec6
Signed-off-by: liamfallon <liam.fallon@est.tech>