Update patch set 1
Patch Set 1: Code-Review-1

The change to the spring version was merged in oparent. This means that some of the changes here are no longer necessary.

We do need to change to the new oparent version, but that will break the build in some components due to checkstyle. We will raise revies for the checkstyle before upgrading oparent here.

Patch-set: 1
Reviewer: Gerrit User 2787 <2787@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1