Add profile for generating Swagger documents
This commit adds a profile for generating Swagger documents and creating
a tarball containing the swagger.json, swagger.html, and swagger.pdf
files.

The profile is triggered in any module when tsts are enabled (In other
words when the skipTests flag is off), and when the following proerty is
set:
        <swagger.generation.phase>post-integration-test</swagger.generation.phase>

This profile assumes that a unit test exists that creates the following
file:
target/swagger/swagger.json

If this file does not exist, the build will fail if the profile is
invoked.

Therefore, to generate the Swagger documentation tarball, a module must
set the swagger.generation.phase propety AND have a unit test that
generates the swagger.json file.

Issue-ID: POLICY-3424
Change-Id: I8ff769e7a28bf0e00969d1bc677fee6908951d8a
Signed-off-by: liamfallon <liam.fallon@est.tech>