Add script to gather data on release process
This script gathers the latest tag version, the master snapshot version,
the number of changed files since the last release, and the docker image
names from each of the Policy Framework repos and stores that data in a
CSV file.

That CSV file is used by other scripts to do release process steps and
to generate commits for the release process.

Issue-ID: POLICY-3835
Change-Id: I2a282286c673bd690b6467cc5617f554f0c6de75
Signed-off-by: liamfallon <liam.fallon@est.tech>