Revert camel version upgrade
Bumping the camel version from 3.7.3 to 3.13.0 breaks the CLAMP
integration tests. The version of camel was bumped due to the log4j
security issue but because the new versions of the log4j libraries are
now managed in the parent POM, the old versions of log4j libs used by
camel will be overridden anyhow. There is no need to step the camel
version.

Issue-ID: POLICY-3862
Change-Id: I2f94e539a7aa4d8b91e4756a052df9bee3f725fd
Signed-off-by: liamfallon <liam.fallon@est.tech>