Release scripts for repos, images, OOM, refs
bumpSnapshots.sh: Steps the patch version of snapshot images based on
the data in the policy release data file.

generateCommit.sh: At the suggestion of OOM, the order of the generated messages is
changed.

releaseRepo.sh: Generates the release yaml file and the commit to
release the maven artifacts of a repo

releaseRepoImages.sh: Generates the release yaml fole and the commit to
release the docker images from a repo

updateOomImages.sh: Generates a commit that updats the references to
Policy Framework docker images in OOM

updateRefs.sh: Updates the parent. common, models., drools-pdp, and
Docker base images in Docker files for reselases or snapshots.

Issue-ID: POLICY-3835
Change-Id: I0766a779c2217a57a33ed37725f11acdfb07a16d
Signed-off-by: liamfallon <liam.fallon@est.tech>