Update docker files base images to snapshots
Currently the policy base images used in docker files in master are hard
revisions of the policy base images. The docker files should use
snapshot policy base images when they are availabe and only change to
released base images at release time.

This commit ensures that the references are updated correctly in Docker
files to snapshots or released base policy images as appropriate

Issue-ID: POLICY-4287
Change-Id: I2c68e1bad6bde2e116c27678c88f8b40dc914b0f
Signed-off-by: liamfallon <liam.fallon@est.tech>