Handle docker refs correctly in policy/docker
The decker references in db-migrator need to be handled specially
because the reference in db-migrator must be set to the new Docker
reference that is to be released rather than the docker reference in the
pf_release_data.csv (which is correct in later phases). This is in
effect a forward reference to the "to-be-released_ docker base image.

Issue-ID: POLICY-4045
Change-Id: Iff3845d06894f7f4c927509fb2fb43f00af3b7ef
Signed-off-by: liamfallon <liam.fallon@est.tech>