Remove policy parent swagger generation profile
This maven profile used the Swagger endpoint to generate the Swagger and
then generated a HTML and PDF version using Asciidoc.

We now have Swagger as a source artifact so this profile is no longer
needed.

This commit als fixes the issue with the example for using this profile
in the documentation, which references a non-existant link.

Issue-ID: POLICY-4431
Change-Id: I970eb392f156662c3bbfb38e978bc46335538925
Signed-off-by: liamfallon <liam.fallon@est.tech>