Update patch set 2
Patch Set 2: Code-Review+1

I think it would be good to add the source document of the png (if it is not already there). 
Otherwise will anyone that wants to update it have to redraw the picture.

Patch-set: 2
Reviewer: Gerrit User 5072 <5072@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1